Created inside romans.js file a function repeatString that mimics String.prototype.repeat()
The replacement is due to the following:
The error message you're encountering indicates that the String.prototype.repeat method is not recognized in your environment. The repeat method is an ES6 feature, and it seems that PhantomJS, being based on an older version of WebKit, does not support this method.

To resolve this issue, you'll need to replace the usage of .repeat() with a function that replicates its behavior using only ES5 features. Here’s a simple polyfill that mimics the functionality of String.prototype.repeat()